+++ admin/invites/invites.controllers.js
@@ -2,7 +2,7 @@ import { PrismaClient } from "@prisma/client";
 import bcrypt from "bcrypt";
 import asyncHandler from "express-async-handler";
 import { v4 as uuidv4 } from "uuid";
-import sendMail from "../services/sendMail.js";
+import sendMail from "../../helper/services/sendMail.js";
 import dotenv from "dotenv";
 dotenv.config()
 const prisma = new PrismaClient();
