+++ helper/middleware/checkTeamExists.js
