+++ helper/middleware/getCurrentWorkspace.js
