+++ hr/employees/employees.routes.js
@@ -1,9 +1,9 @@
 import express from "express";
-import { verifyToken } from "../middleware/authenticate.js";
+import { verifyToken } from "../../helper/middleware/authenticate.js";
 import {
   getCurrentEmployee,
   getCurrentWorkspace,
-} from "../middleware/index.js";
+} from "../../helper/middleware/index.js";
 import {
   changeEmployeeRole,
   deleteEmployee,
