+++ prisma/schema.prisma
@@ -1,4 +1,4 @@
-//🧩Don't Edit this file.✨Generated in Mon Dec 18 2023 09:37:45 GMT+0100 (West Africa Standard Time)✨
+//🧩Don't Edit this file.✨Generated in Tue Dec 19 2023 20:03:32 GMT+0100 (West Africa Standard Time)✨
 
 generator client {
   provider      = "prisma-client-js"
@@ -88,7 +88,7 @@ enum UserRole {
 }
 
 // Project Management 
-
+// TODO: have priority and status model with thier colors but now in a json format where organanization can several status type and colors and assign them to a particular project 
 model Space {
   id    String @id @default(cuid())
   title String
@@ -483,6 +483,8 @@ model Team {
   name  String  @unique
   alias String?
 
+  role TeamRole @default(MEMBER)
+
   workspaceId  String
   workspace    Workspace      @relation(fields: [workspaceId], references: [id], onDelete: Cascade)
   employeeTeam EmployeeTeam[]
@@ -549,12 +551,12 @@ model Leave {
   id String @id @default(cuid())
 
   appliedDate   DateTime       @default(now())
-  startDate     DateTime
-  endDate       DateTime
-  acceptedDate  DateTime?
+  startDate     DateTime       @db.Date
+  endDate       DateTime       @db.Date
   duration      Int
-  reason        String?
-  leaveType     LeaveType[]
+  reason        String
+  leaveTypeId   String
+  leaveType     LeaveType      @relation(fields: [leaveTypeId], references: [id])
   reviewComment LeaveComment[]
   status        LeaveStatus    @default(PENDING)
 
@@ -565,24 +567,19 @@ model Leave {
   updatedAt DateTime @updatedAt
 
   @@index([employeeId])
+  @@index([leaveTypeId])
 }
 
-model Education {
-  id String @id @default(cuid())
+model LeaveType {
+  id   String @id @default(cuid())
+  name String
 
-  institution   String?
-  courseOfStudy String?
-  result        String?
-  degree        String?
-  startDate     DateTime?
-  endDate       DateTime?
-  status        Boolean   @default(true)
+  leaves Leave[]
 
   employeeId String
-  employee   Employee @relation(fields: [employeeId], references: [id], onDelete: Cascade)
-
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
+  employee   Employee @relation(fields: [employeeId], references: [id])
+  createdAt  DateTime @default(now())
+  updatedAt  DateTime @updatedAt
 
   @@index([employeeId])
 }
@@ -605,21 +602,24 @@ model LeaveComment {
   @@index([leaveId])
 }
 
-model LeaveType {
-  id    String @id @default(cuid())
-  title String
+model Education {
+  id String @id @default(cuid())
+
+  institution   String?
+  courseOfStudy String?
+  result        String?
+  degree        String?
+  startDate     DateTime?
+  endDate       DateTime?
+  status        Boolean   @default(true)
 
   employeeId String
-  employee   Employee @relation(fields: [employeeId], references: [id])
-
-  leave   Leave?  @relation(fields: [leaveId], references: [id])
-  leaveId String?
+  employee   Employee @relation(fields: [employeeId], references: [id], onDelete: Cascade)
 
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 
   @@index([employeeId])
-  @@index([leaveId])
 }
 
 model PerformanceAppraisal {
